+++ packages/eslint-config-dawn/ts.js
@@ -1,4 +1,4 @@
 // Alias for typescript.js
 module.exports = {
-  extends: require.resolve('./typescript'),
+  extends: [require.resolve('./typescript')],
 };