+++ packages/middleware-lint/src/test.js
@@ -1,7 +1,0 @@
-console.log('你好', {
-  a: 0,
-  b: 1,
-});
-const exp = 'dsafDAdsaf';
-console.log(exp);
-// const bar_foo = 2;