+++ packages/middleware-lint/src/utils/debug.ts
@@ -1,0 +1,3 @@
+import createDebugger from "debug";
+
+export default createDebugger("dn:middleware:lint");