+++ packages/middleware-lint/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.3](https://github.com/alibaba/dawn/compare/@dawnjs/dn-middleware-lint@3.1.2...@dawnjs/dn-middleware-lint@3.1.3) (2021-11-12)
+
+### Bug Fixes
+
+- only install deps if needed ([111cd65](https://github.com/alibaba/dawn/commit/111cd65713ef211df8e3ab3ab1e3302f9be6994d))
+
+## [3.1.2](https://github.com/alibaba/dawn/compare/@dawnjs/dn-middleware-lint@3.1.1...@dawnjs/dn-middleware-lint@3.1.2) (2021-11-12)
+
+### Bug Fixes
+
+- fix save dev flag ([7797fee](https://github.com/alibaba/dawn/commit/7797fee6b070620f52a173e5dedca83b9c32728b))
+
+## [3.1.1](https://github.com/alibaba/dawn/compare/@dawnjs/dn-middleware-lint@3.1.0...@dawnjs/dn-middleware-lint@3.1.1) (2021-11-12)
+
+### Bug Fixes
+
+- fix readFile issue ([9842312](https://github.com/alibaba/dawn/commit/984231235e8b0f08c07bcb69a48c629dfff1e3f9))
+
+# [3.1.0](https://github.com/alibaba/dawn/compare/@dawnjs/dn-middleware-lint@3.0.12...@dawnjs/dn-middleware-lint@3.1.0) (2021-11-12)
+
+### Bug Fixes
+
+- adjust deps logic ([1836bb3](https://github.com/alibaba/dawn/commit/1836bb3ee5f1421935ab97f4d3102f532f90c6a6))
+
+### Features
+
+- postinstall required deps ([3c36b77](https://github.com/alibaba/dawn/commit/3c36b7733396952c89e887321f5d9f3d89641df3))
+
 ## 3.0.12 (2021-06-16)
 
 **Note:** Version bump only for package @dawnjs/dn-middleware-lint